Fix c2c2c_identity test
Reusing a plan for different arrays than it was created with requires that they
have the same size and alignment as was used for creating the plan. While the
size of `a` and `b` is identical, there is no such guarantee for their
alignment. Therefore one may not simply swap them.

Fix this by using separate plans for the `in: a, out: b` and `in: b, out: a`
invocations.