Fix r2r2r_identity test
Reusing a plan for different arrays than it was created with requires that they
have the same size and alignment as was used for creating the plan. By
assigning an entirely new vec to `a` - which may have a different alignment -
this test currently violating that requirement.

Fix this by filling the existing vec instead of allocating a new one